requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
Werkzeug==0.9.6
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
boto==2.29.1
celery==3.1.12
dj-database-url==0.3.0
dj-static==0.0.5
django-admin-export==2.0
django-ajax-selects==1.3.4
django-apptemplates==0.0.1
django-bootstrap==0.2.4
django-bower==5.0.1
django-cached-field==1.2.6
django-celery==3.1.10
django-ckeditor==4.0.2
django-countries==2.1.2
django-crispy-forms==1.4.0
django-custom-field==2.2
django-dajax==0.9.2
django-dajaxice==0.6
-e git://github.com/tzulberti/django-datefilterspec.git@c9ed941b9022f3dcbd2c6f83a4be731544eca81e#egg=django_daterange_filter-master
django-extensions==1.3.8
django-filter==0.7
django-floppy-gumby==0.1
django-floppyforms==1.2.0
django-grappelli==2.5.3
django-ldap-groups==0.1.3
django-localflavor==1.0
django-mass-edit==2.2
django-pagination==1.0.7
django-positions==0.5.0
-e git://github.com/burke-software/django-report-builder.git@ae4da2e5e28a92bbaf7114606d768e7547ffa789#egg=django_report_builder-dev
-e git://github.com/burke-software/django-report-scaffold.git@d9070e7cc14f0f40a7fec3ca9d9e85cd07157711#egg=django_report_scaffold-origin/master
django-report-utils==0.2.5
-e git://github.com/burke-software/django-responsive-dashboard.git@5ccc205f1a3b7fd1913c26c8cc30c54eaac4312e#egg=django_responsive_dashboard-dev
django-reversion==1.8.1
django-rosetta==0.7.4
-e git://github.com/zsiciarz/django-rosetta-grappelli.git@d2f8b26ecb12ed369c8e496d6a8f401e923745d0#egg=django_rosetta_grappelli-origin/develop
-e git://github.com/burke-software/django-simple-import.git@34e5bf62a3d26a491056c479706fc5880548f31e#egg=django_simple_import-dev
django-storages==1.1.8
django-su==0.4.5
django-toolbelt==0.0.1
django-widget-tweaks==1.3
django-xadmin==0.5.0
elementtree==1.2.7-20070827-preview
feedparser==5.1.3
gunicorn==19.0.0
httpagentparser==1.7.3
jdcal==1.0
kombu==3.0.19
odfpy==0.9.6
openpyxl==2.0.3
polib==1.0.4
poster==0.8.1
psycopg2==2.5.3
pystache==0.5.4
python-dateutil==2.2
pytz==2014.4
raven==5.0.0
reportlab==3.1.8
requests==2.3.0
simplejson==3.5.2
six==1.7.2
static==1.0.2
suds==0.4
wsgiref==0.1.2
xlrd==0.9.3
xlwt==0.7.5
